+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## UNRELEASED
 
+FEATURES:
+* config: Map host devices into container. [[GH-41](https://github.com/hashicorp/nomad-driver-podman/pull/41)]
+* config: Stream logs via API, support journald log driver. [[GH-99](https://github.com/hashicorp/nomad-driver-podman/pull/99)]
+
+BUG FIXES:
+* log: Use error key context to log errors rather than Go err style. [[GH-126](https://github.com/hashicorp/nomad-driver-podman/pull/126)]
+
 ## 0.3.0
 
 * config: Image registry authentication [[GH-71](https://github.com/hashicorp/nomad-driver-podman/issues/71)]